+++ lang/en/theme_campus.php
@@ -405,7 +405,7 @@ $string['coursecategoryresponsivebackgroundimagedesc'] = 'Please upload your cus
 $string['imageleft'] = 'Left';
 $string['imageright'] = 'Right';
 
-// Forum header.
+// Forum header colour.
 $string['forumheadertextcolour'] = 'Forum header text colour';
 $string['forumheadertextcolourdesc'] = 'Set the forum header text colour.';
 
@@ -438,3 +438,11 @@ $string['socialnetworkicon_desc'] = 'Social network icon number ';
 // admin_setting_configinteger.
 $string['asconfigintlower'] = '{$a->value} is less than the lower range limit of {$a->lower}';
 $string['asconfigintupper'] = '{$a->value} is greater than the upper range limit of {$a->upper}';
+
+// Forum.
+$string['forumsettings'] = 'Forum general';
+$string['forumsettingsdesc'] = 'Configure the general forum settings for the theme here.';
+$string['forumhtmlemailheader'] = 'Forum html email header';
+$string['forumhtmlemailheaderdesc'] = 'Configure the forum html email header for the theme here.';
+$string['forumhtmlemailfooter'] = 'Forum html email footer';
+$string['forumhtmlemailfooterdesc'] = 'Configure the forum html email footer for the theme here.';